+++ R/crayon.R
@@ -1,5 +1,5 @@
-# message <- function(...,fg="silver",bg="bgWhite")
+# message <- function(...,fg="black",bg="bgWhite")
 # {
-#   style <- crayon::combine_styles(fg,bg)
+#   style <- crayon::combine_styles(fg)
 #   base::message(style(...))
 # }